--- src/main/java/kr/hhplus/be/server/api/common/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.api.common.exception;
-
-public record ErrorResponse(
-
-) {
-}
