+++ src/main/java/kr/hhplus/be/server/api/reservation/presentation/ReservationController.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import kr.hhplus.be.server.api.common.response.ApiResponse;
+import kr.hhplus.be.server.common.response.ApiResponse;
 import kr.hhplus.be.server.api.reservation.application.dto.ReservationResult;
 import kr.hhplus.be.server.api.reservation.presentation.dto.ReservationRequest;
 import kr.hhplus.be.server.api.reservation.application.ReservationFacade;
