+++ src/main/java/kr/hhplus/be/server/api/user/domain/repository/UserRepository.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.api.user.domain.repository;
 
 import jakarta.persistence.LockModeType;
-import kr.hhplus.be.server.api.concert.domain.entity.ConcertSeat;
 import kr.hhplus.be.server.api.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
