+++ src/main/java/kr/hhplus/be/server/common/aop/lock/RedisLockAop.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.api.common.aop.lock;
+package kr.hhplus.be.server.common.aop.lock;
 
-import kr.hhplus.be.server.api.common.exception.CustomException;
-import kr.hhplus.be.server.api.common.exception.ErrorCode;
+import kr.hhplus.be.server.common.exception.CustomException;
+import kr.hhplus.be.server.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -10,7 +10,6 @@ import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.reflect.MethodSignature;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
-import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 import java.lang.reflect.Method;
@@ -19,14 +18,14 @@ import java.lang.reflect.Method;
 @RequiredArgsConstructor
 @Component
 @Aspect
-@Order(0) //트랜잭션보다 우선 실행
+//@Order(0) //트랜잭션보다 우선 실행
 public class RedisLockAop {
     private static final String REDIS_LOCK_PREFIX = "lock:";
 
     private final RedissonClient redissonClient;
     private final AopForTransaction transaction;
 
-    @Around("@annotation(kr.hhplus.be.server.api.common.aop.lock.RedisLock)")
+    @Around("@annotation(kr.hhplus.be.server.common.aop.lock.RedisLock)")
     public Object getLock(ProceedingJoinPoint joinPoint) throws Throwable {
         MethodSignature signature = (MethodSignature) joinPoint.getSignature();
         Method method = signature.getMethod();
