+++ src/main/java/kr/hhplus/be/server/common/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.common.exception;
+
+public record ErrorResponse(
+
+) {
+}
