+++ src/test/java/kr/hhplus/be/server/api/concert/ConcertServiceTest.java
@@ -16,7 +16,7 @@ import kr.hhplus.be.server.api.concert.domain.entity.ConcertSchedule;
 import kr.hhplus.be.server.api.concert.domain.repository.ConcertScheduleRepository;
 import kr.hhplus.be.server.api.concert.domain.entity.ConcertSeat;
 import kr.hhplus.be.server.api.concert.domain.repository.ConcertSeatRepository;
-import kr.hhplus.be.server.api.common.exception.ErrorCode;
+import kr.hhplus.be.server.common.exception.ErrorCode;
 import org.junit.Test;
 
 //@ExtendWith(MockitoExtension.class)
