+++ src/test/java/kr/hhplus/be/server/api/reservation/ReservationFacadeIntegrationTest.java
@@ -6,19 +6,22 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDateTime;
+import java.util.List;
 
 import kr.hhplus.be.server.api.concert.domain.entity.ConcertSeat;
 import kr.hhplus.be.server.api.concert.domain.repository.ConcertSeatRepository;
 import kr.hhplus.be.server.api.reservation.application.ReservationFacade;
 import kr.hhplus.be.server.api.reservation.application.dto.ReservationDto;
 import kr.hhplus.be.server.api.reservation.application.dto.ReservationResult;
+import kr.hhplus.be.server.api.reservation.domain.entity.Reservation;
+import kr.hhplus.be.server.api.reservation.domain.repository.ReservationRepository;
 import kr.hhplus.be.server.api.token.domain.entity.Token;
 import kr.hhplus.be.server.api.token.domain.repository.TokenRepository;
 import kr.hhplus.be.server.api.user.application.dto.BalanceHistoryDto;
 import kr.hhplus.be.server.api.user.domain.entity.BalanceHistoryType;
 import kr.hhplus.be.server.api.user.domain.entity.User;
 import kr.hhplus.be.server.api.user.domain.repository.UserRepository;
-import kr.hhplus.be.server.api.common.exception.ErrorCode;
+import kr.hhplus.be.server.common.exception.ErrorCode;
 import kr.hhplus.be.server.api.user.domain.service.UserService;
 import kr.hhplus.be.server.util.DatabaseCleaner;
 import org.junit.jupiter.api.BeforeEach;
@@ -39,6 +42,8 @@ class ReservationFacadeIntegrationTest {
 	private UserRepository userRepository;
     @Autowired
     private ConcertSeatRepository concertSeatRepository;
+	@Autowired
+	private ReservationRepository reservationRepository;
     @Autowired
     private TokenRepository tokenRepository;
 	@Autowired
@@ -68,18 +73,23 @@ class ReservationFacadeIntegrationTest {
 		assertThat(sut.isReserved()).isEqualTo(true);
 		assertThat(sut.seatNo()).isEqualTo(dto.seatNo());
 		assertThat(sut.expiredAt()).isNotNull();
+		List<Reservation> list = reservationRepository.findAll();
+		assertThat(list).hasSize(1);
 	}
 
 	@Test
 	void 임시_예약_이미예약된좌석일시_ALREADY_RESERVED_SEAT() {
 		// given
-		ReservationDto dto = new ReservationDto(1L, 1L, 1L, 11, mockTime.toLocalDate(), BigDecimal.valueOf(55000));
+		ReservationDto successRes = new ReservationDto(1L, 1L, 1L, 10, mockTime.toLocalDate(), BigDecimal.valueOf(55000));
+		ReservationDto failRes = new ReservationDto(1L, 2L, 1L, 10, mockTime.toLocalDate(), BigDecimal.valueOf(55000));
 
 		// when
-		reservationFacade.makeReservation(dto);
+		reservationFacade.makeReservation(successRes);
 
 		// then
-		assertThatThrownBy(() -> reservationFacade.makeReservation(dto)).hasMessage(ErrorCode.ALREADY_RESERVED_SEAT.getReason());
+		assertThatThrownBy(() -> reservationFacade.makeReservation(failRes)).hasMessage(ErrorCode.ALREADY_RESERVED_SEAT.getReason());
+		Reservation sut = reservationRepository.findBySeatIdAndIsReservedTrue(1L);
+		assertThat(sut.getUserId()).isEqualTo(1L);
 	}
 
 	@Test
